audio_evals/lib/whisper/requirements.txt
accelerate>=0.26.0
openai-whisper
scipy
soundfile
torchaudio==2.5.1
transformers==4.49.0
